refactor(driver): remove InsertValues legacy load path from all drivers
Deletes Driver.InsertValues from the dispatcher interface and from
all five concrete drivers (postgres / mysql / picodata / ydb / noop).
Removes the supporting InsertDescriptor-based pipeline:
pkg/driver/sqldriver/insert.go (InsertPlainQuery/InsertPlainBulk) plus
pkg/driver/sqldriver/queries/{builder,insert_query,query_common,generators}.go.

postgres/driver_native.go (streamingCopySource) and the legacy
insertValuesNative method on ydb go with them — BulkUpsert stays via
toYDBValue which insert_spec.go still consumes.

Tests: postgres/driver_test.go and the picodata InsertValues tests
are deleted; a new postgres/test_helpers_test.go preserves the
newTestDriver + mockExecutor helpers used by rows_test.go.
