+++ pkg/driver/postgres/driver.go
@@ -2,7 +2,6 @@ package postgres
 
 import (
 	"context"
-	"fmt"
 	"time"
 
 	"github.com/jackc/pgx/v5"
@@ -10,14 +9,11 @@ import (
 	"github.com/jackc/pgx/v5/pgxpool"
 	"go.uber.org/zap"
 
-	"github.com/stroppy-io/stroppy/pkg/common/generate"
 	"github.com/stroppy-io/stroppy/pkg/common/logger"
 	stroppy "github.com/stroppy-io/stroppy/pkg/common/proto/stroppy"
 	"github.com/stroppy-io/stroppy/pkg/driver"
 	"github.com/stroppy-io/stroppy/pkg/driver/postgres/pool"
 	"github.com/stroppy-io/stroppy/pkg/driver/sqldriver"
-	sqlqueries "github.com/stroppy-io/stroppy/pkg/driver/sqldriver/queries"
-	"github.com/stroppy-io/stroppy/pkg/driver/stats"
 )
 
 const dbConnectionTimeout = 5 * time.Second
@@ -146,52 +142,3 @@ func (d *Driver) RunQuery(
 	return sqldriver.RunQuery(ctx, d.pool, NewRows, PgxDialect{}, d.logger, sql, args)
 }
 
-// InsertValues inserts multiple rows into the database based on the descriptor.
-// It supports three methods:
-// - PLAIN_QUERY: executes individual INSERT statements for each row
-// - PLAIN_BULK: executes batched bulk INSERT statements using multi-row VALUES syntax
-// - NATIVE: uses PostgreSQL's COPY protocol for fast bulk insertion.
-func (d *Driver) InsertValues(
-	ctx context.Context,
-	descriptor *stroppy.InsertDescriptor,
-) (*stats.Query, error) {
-	builder, err := sqlqueries.NewQueryBuilder(
-		d.logger,
-		PgxDialect{},
-		generate.ResolveSeed(descriptor.GetSeed()),
-		descriptor,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("can't create query builder: %w", err)
-	}
-
-	switch descriptor.GetMethod() {
-	case stroppy.InsertMethod_PLAIN_QUERY:
-		return sqldriver.InsertPlainQuery(ctx, d.pool, builder)
-	case stroppy.InsertMethod_PLAIN_BULK:
-		return sqldriver.InsertPlainBulk(ctx, d.pool, builder, d.bulkSize)
-	case stroppy.InsertMethod_NATIVE:
-		return d.insertValuesNative(ctx, builder)
-	default:
-		d.logger.Panic("unexpected proto.InsertMethod")
-
-		return nil, nil //nolint:nilnil // unreachable after panic
-	}
-}
-
-// insertValuesNative uses PostgreSQL's COPY protocol for fast bulk insertion.
-// It streams values on-demand without loading all rows into memory.
-func (d *Driver) insertValuesNative(
-	ctx context.Context,
-	builder *sqlqueries.QueryBuilder,
-) (*stats.Query, error) {
-	cols := builder.Columns()
-	stream := newStreamingCopySource(builder)
-	start := time.Now()
-
-	if _, err := d.pool.CopyFrom(ctx, pgx.Identifier{builder.TableName()}, cols, stream); err != nil {
-		return nil, err
-	}
-
-	return &stats.Query{Elapsed: time.Since(start)}, nil
-}
