--- pkg/driver/postgres/driver_test.go
@@ -1,194 +1,0 @@
-package postgres
-
-import (
-	"context"
-	"testing"
-
-	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/pgconn"
-	"github.com/jackc/pgx/v5/pgxpool"
-	"github.com/pashagolub/pgxmock/v4"
-	"github.com/stretchr/testify/require"
-
-	"github.com/stroppy-io/stroppy/pkg/common/logger"
-	stroppy "github.com/stroppy-io/stroppy/pkg/common/proto/stroppy"
-)
-
-func ptr[T any](v T) *T {
-	return &v
-}
-
-// mockExecutor wraps pgxmock.PgxPoolIface to satisfy the Executor interface
-// by adding ExecContext/QueryContext shims (which delegate to Exec/Query).
-type mockExecutor struct {
-	pgxmock.PgxPoolIface
-}
-
-func (m *mockExecutor) ExecContext(ctx context.Context, sql string, args ...any) (pgconn.CommandTag, error) {
-	return m.Exec(ctx, sql, args...)
-}
-
-func (m *mockExecutor) QueryContext(ctx context.Context, sql string, args ...any) (pgx.Rows, error) {
-	return m.Query(ctx, sql, args...)
-}
-
-func (m *mockExecutor) Config() *pgxpool.Config { return nil }
-
-type testDriver struct {
-	*Driver
-}
-
-func newTestDriver(mockPool pgxmock.PgxPoolIface) *testDriver {
-	return &testDriver{
-		Driver: &Driver{
-			logger: logger.Global(),
-			pool:   &mockExecutor{mockPool},
-		},
-	}
-}
-
-func TestDriver_InsertValuesPlainQuery(t *testing.T) {
-	mock, err := pgxmock.NewPool()
-	require.NoError(t, err)
-
-	defer mock.Close()
-
-	drv := newTestDriver(mock)
-
-	ctx := context.Background()
-	descriptor := &stroppy.InsertDescriptor{
-		Count:     3,
-		TableName: "test_table",
-		Method:    stroppy.InsertMethod_PLAIN_QUERY.Enum(),
-		Params: []*stroppy.QueryParamDescriptor{
-			{
-				Name: "id",
-				GenerationRule: &stroppy.Generation_Rule{
-					Kind: &stroppy.Generation_Rule_Int64Range{
-						Int64Range: &stroppy.Generation_Range_Int64{
-							Min: ptr[int64](1),
-							Max: 100,
-						},
-					},
-					Unique: ptr(true),
-				},
-			},
-		},
-	}
-
-	// Expect 3 insert executions
-	for range descriptor.GetCount() {
-		mock.ExpectExec("INSERT INTO test_table").
-			WithArgs(pgxmock.AnyArg()).
-			WillReturnResult(pgxmock.NewResult("INSERT", 1))
-	}
-
-	stats, err := drv.InsertValues(ctx, descriptor)
-	require.NoError(t, err)
-	require.NotNil(t, stats)
-
-	require.NoError(t, mock.ExpectationsWereMet())
-}
-
-func TestDriver_InsertValuesNative(t *testing.T) {
-	mock, err := pgxmock.NewPool()
-	require.NoError(t, err)
-
-	defer mock.Close()
-
-	drv := newTestDriver(mock)
-
-	ctx := context.Background()
-	descriptor := &stroppy.InsertDescriptor{
-		Count:     5,
-		TableName: "test_table",
-		Method:    stroppy.InsertMethod_NATIVE.Enum(),
-		Params: []*stroppy.QueryParamDescriptor{
-			{
-				Name: "id",
-				GenerationRule: &stroppy.Generation_Rule{
-					Kind: &stroppy.Generation_Rule_Int64Range{
-						Int64Range: &stroppy.Generation_Range_Int64{
-							Min: ptr[int64](1),
-							Max: 100,
-						},
-					},
-					Unique: ptr(true),
-				},
-			},
-			{
-				Name: "name",
-				GenerationRule: &stroppy.Generation_Rule{
-					Kind: &stroppy.Generation_Rule_StringConst{
-						StringConst: "test_name",
-					},
-				},
-			},
-		},
-	}
-
-	// Expect one CopyFrom call with 5 rows
-	mock.ExpectCopyFrom(
-		[]string{"test_table"},
-		[]string{"id", "name"},
-	).WillReturnResult(int64(descriptor.GetCount()))
-
-	stats, err := drv.InsertValues(ctx, descriptor)
-	require.NoError(t, err)
-	require.NotNil(t, stats)
-
-	require.NoError(t, mock.ExpectationsWereMet())
-}
-
-func TestDriver_InsertValuesNativeLargeBatch(t *testing.T) {
-	mock, err := pgxmock.NewPool()
-	require.NoError(t, err)
-
-	defer mock.Close()
-
-	drv := newTestDriver(mock)
-
-	ctx := context.Background()
-	descriptor := &stroppy.InsertDescriptor{
-		Count:     10000,
-		TableName: "test_table",
-		Method:    stroppy.InsertMethod_NATIVE.Enum(),
-		Params: []*stroppy.QueryParamDescriptor{
-			{
-				Name: "id",
-				GenerationRule: &stroppy.Generation_Rule{
-					Kind: &stroppy.Generation_Rule_Int64Range{
-						Int64Range: &stroppy.Generation_Range_Int64{
-							Min: ptr[int64](1),
-							Max: 1000000,
-						},
-					},
-					Unique: ptr(true),
-				},
-			},
-			{
-				Name: "value",
-				GenerationRule: &stroppy.Generation_Rule{
-					Kind: &stroppy.Generation_Rule_Int64Range{
-						Int64Range: &stroppy.Generation_Range_Int64{
-							Min: ptr[int64](1),
-							Max: 1000,
-						},
-					},
-				},
-			},
-		},
-	}
-
-	// Expect one CopyFrom call with 10000 rows - demonstrates streaming without memory issues
-	mock.ExpectCopyFrom(
-		[]string{"test_table"},
-		[]string{"id", "value"},
-	).WillReturnResult(int64(descriptor.Count))
-
-	stats, err := drv.InsertValues(ctx, descriptor)
-	require.NoError(t, err)
-	require.NotNil(t, stats)
-
-	require.NoError(t, mock.ExpectationsWereMet())
-}
